refactor(network): move to converters-starknetapi Header & Signature to protobuf SignedBlockHeader
